+++ lib/e621.dart
@@ -1,6 +1,9 @@
 /// A wrapper for the e621 api, complete with models that handle JSON responses and built-in rate-limiting.
 //@TestOn()
+@Deprecated("Use 'e621-api package'")
 library;
 
+@Deprecated("Use 'e621-api package'")
 export 'e621_api.dart';
+@Deprecated("Use 'e621-api package'")
 export 'e621_models.dart';