+++ lib/e621_api.dart
@@ -1,9 +1,76 @@
 /// A wrapper for the e621 api, with only the API calling code, with built-in rate-limiting.
 //@TestOn()
+@Deprecated("Use 'e621-api package'")
 library;
 
-export 'src/e621/e621.dart';
+@Deprecated("Use 'e621-api package'")
+export 'src/e621/e621.dart'
+    /* show
+        activeCredentials,
+        alphabetLowerAndUpper,
+        baseUri,
+        baseUriNsfw,
+        baseUriSfw,
+        disallowedInTagName,
+        e6ValidMetaTagCharacters,
+        hostNameNsfw,
+        hostNameSfw,
+        maxPageNumber,
+        maxPostsPerSearch,
+        maxPostsPerSearchByPageNumber,
+        maxTagsPerSearch,
+        numerals,
+        useNsfw,
+        doesPostUpdateHaveChanges,
+        foldIterableForUrl,
+        generateDiff,
+        getPageString,
+        initAddToSetRequest,
+        initCreateFavoriteRequest,
+        initCreateNoteRequest,
+        initCreatePoolRequest,
+        initCreateSetRequest,
+        initDbExportRequest,
+        initDeleteFavoriteRequest,
+        initDeleteNoteRequest,
+        initGetPoolRequest,
+        initGetPostRequest,
+        initGetSetRequest,
+        initGetUserRequest,
+        initListFavoritesRequest,
+        initListFavoritesWithCredentialsRequest,
+        initListFavoritesWithIdRequest,
+        initPostCastVoteRequest,
+        initPostSearchRequest,
+        initPostSearchRequestChecked,
+        initRemoveFromSetRequest,
+        initRevertNoteRequest,
+        initRevertPoolRequest,
+        initSearchNotesRequest,
+        initSearchPoolsRequest,
+        initSearchPopularRequest,
+        initSearchPopularRequestUnconstrained,
+        initSearchSetsRequest,
+        initSearchTagAliasesRequest,
+        initSearchTagImplicationsRequest,
+        initSearchTagsRequest,
+        initSearchUsersRequest,
+        initUpdatePoolRequest,
+        initUpdatePostRequest,
+        initUpdateSetPostsRequest,
+        initUpdateSetRequest,
+        initVotePostRequest,
+        initWikiGetPageRequest,
+        initWikiSearchRequest,
+        validateCredentials,
+        Endpoint,
+        initSearchPostRequest */
+    ;
+@Deprecated("Use 'e621-api package'")
 export 'src/e621/credentials.dart';
+@Deprecated("Use 'e621-api package'")
 export 'src/e621/middleware.dart';
+@Deprecated("Use 'e621-api package'")
 export 'src/e621/search_enums.dart';
+@Deprecated("Use 'e621-api package'")
 export 'src/e621/general_enums.dart';